rename to db_extract
